services/whatsapp-ingestor/requirements.txt
google-cloud-pubsub==2.23.0
fastapi==0.111.0
uvicorn==0.29.0
